requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pypdf==4.0.1
python-docx==1.1.0
httpx==0.26.0
python-dotenv==1.0.0
